need to optimize place_val

+++ srcs/utils/utils.h
@@ -6,7 +6,7 @@
 /*   By: lejulien <lejulien@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/05 13:06:31 by lejulien          #+#    #+#             */
-/*   Updated: 2021/03/28 14:59:23 by lejulien         ###   ########.fr       */
+/*   Updated: 2021/03/29 17:40:57 by lejulien         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,5 +16,5 @@
 int		ft_atoi(const char *str);
 void	ft_putnbr_fd(int nb, int fd);
 void	ft_putstr(const char *str);
-int		ft_sqrt(int nb);
+float	ft_sqrt(float nb);
 #endif
